Allow for the creation of Magnitude.
There's an unsafe portion to this function: rather than require Default
on T, we just utilize MaybeUninit. We therefore need to keep track of
the number of values that are truly initialized. This would still
probably have to be done if we did use Default -- there's no guarantee
that Default values are the smallest possible.