Panic if we haven't initialized enough values.
We're likely going to have to develop our own `IntoIter` struct;
we need to be able to slice down the underlying buf and that would
require giving `core::array:IntoIter` the boot. Another problem is
the sorting: the solution will likely to be to enforce the type
of sorting we do in the insertion when `num_initialized` > N, but
alter the range of values that we're willing to compare to. We'd
find the index and then slide everything above it up.