Test for the uninitialized case.
Of note, when the BPQ is uninitialized the underlying buf hasn't been
sorted; we don't need to do the `rev`. This probably isn't a good thing;
the sorted behavior should be the same regardless of the number of
initialized values.

To fix this, we'd have to mess with the index calculation/insertion
within the BPQ structure.