Cartesian Product over slices.
When computing DST combinations we want to find the intersection
over all combining elements. This introduces the CartesianProduct
and a relevant function `inc` to advance over each element.

1. We don't utilize iterators. There are iterator implementations
(see itertools), but we're likely to be passed in a slice of mem
anyway.

2. This function is recursive which might not pass for safety-critical
programs; at some point, we should rewrite it such that it is iterative.