Fill out more of the EWAF.
The choice of eta is difficult -- will need to read more of PLG.
Moreover, there's an issue here with the calculation of `exp`;
there's no exp function within the core lib. I assume I have to
utilize a HAL on the microcontroller (or something to that regard)?
So, ideally, we're injecting an `exp()` into this struct that
at compile-time gets figured out.

More evidence that I need to build out some builder-type structure
for configuring these forecasters.