A sine wave test.
The work for `eta` needs to be hammered out -- I think it'd be smart
to hammer out a builder-type API and then use that to select the
eta parameter.

Additionally, this commit reveals that the program is bugged --
specifically the weight updating code (we're getting a NaN after the
second iteration).