Support calculations of plausability.
Involves the computation of `not`. Here's a problem though: mass
values are a generic `T`; this isn't an issue for `bel` because we
can require an implementation of `sum`; for `pl`, we need a value
of '1' to compute `1 - bel(not(q))`. At the moment I've required `T`
to have `From<u8>` which allows us to convert from `1u8`.

This isn't great, but I think it's a workable solution.

Introduces tests for `pl` and refactors all the tests to be a
bit cleaner. The interval `not` needs to be worked out.