check-in; refactor of the HashMap code.
The major thing here is that it *might* be faster to just dump the
hashing and search a linear arr. It's something that would be worth
benchmarking.

ARX might be better than FNV to remove the mul. The idea of a
`scheme` might need to be worked -- we're passing it a `[Option; N2]`
arr; it might be better to require that these schemes to be passed
an iterator?