HashMap cleanups.
Rename to `SummationHashMap`, placing it within its own module.

Move the `Copy` bound to the `insert` method; the logic behind this is
that `MulAssign` will be used in every `insert` type operation, but this
specific impl depends on `Copy` -- we could benchmark this later. `MulAssign`
will be needed regardless.

Additional documentation added, additional tests added.

Moving forward -- this isn't a hashmap, should it be? Seems like an avenue to bench.